refactor(ui): shared status + format utils for teacher/student
Two status vocabularies (activity: active/idle/not_started; submission:
graded/submitted/auto_submitted/late/missing/pending) and the fmtDate/fmtSize/
downloadBlob helpers were re-implemented inline across ~10 components and had
started to drift. Centralised into $lib/utils/status.ts and $lib/utils/format.ts
and imported everywhere.

fmtDate kept its two real semantics as fmtDate (date only) and fmtDateTime
(date + time) so no call site changes behaviour. The one visible tidy-up: the
teacher assignment-detail submission badges now use the same Title-Case labels
as the student view (existing i18n keys), instead of raw lowercase status keys.

Prettier + eslint clean, vitest green; no type errors introduced.
