+++ ui/src/lib/components/student/pages/Assignments.svelte
@@ -8,6 +8,11 @@
 		downloadMySubmissionAttachment,
 		type StudentAssignment
 	} from '$lib/apis/assignments';
+	import {
+		submissionStatusStyle as statusStyle,
+		submissionStatusLabel as statusLabel
+	} from '$lib/utils/status';
+	import { fmtDate, downloadBlob as saveBlob } from '$lib/utils/format';
 	import { uploadFile } from '$lib/apis/files';
 	import { getExam, startExam, type ExamInfo, type ExamConfig } from '$lib/apis/exams';
 	import ExamShell from '$lib/components/student/ExamShell.svelte';
@@ -57,14 +62,6 @@
 		const input = e.target as HTMLInputElement;
 		file = input.files && input.files.length ? input.files[0] : null;
 	}
-	function saveBlob(blob: Blob, filename: string) {
-		const url = URL.createObjectURL(blob);
-		const a = document.createElement('a');
-		a.href = url;
-		a.download = filename;
-		a.click();
-		URL.revokeObjectURL(url);
-	}
 	async function getTeacherFile() {
 		if (!active?.attachment_id) return;
 		try {
@@ -86,27 +83,6 @@
 		} catch (err) {
 			/* ignore */
 		}
-	}
-
-	const statusStyle: Record<string, string> = {
-		graded: 'bg-blue-100 text-blue-700 dark:bg-blue-900/30 dark:text-blue-300',
-		submitted: 'bg-green-100 text-green-700 dark:bg-green-900/30 dark:text-green-300',
-		auto_submitted: 'bg-yellow-100 text-yellow-700 dark:bg-yellow-900/30 dark:text-yellow-300',
-		late: 'bg-amber-100 text-amber-700 dark:bg-amber-900/30 dark:text-amber-300',
-		missing: 'bg-red-100 text-red-700 dark:bg-red-900/30 dark:text-red-300',
-		pending: 'bg-gray-100 text-gray-500 dark:bg-gray-700 dark:text-gray-400'
-	};
-	const statusLabel: Record<string, string> = {
-		graded: 'Graded',
-		submitted: 'Submitted',
-		auto_submitted: 'Auto-submitted',
-		late: 'Submitted late',
-		missing: 'Missing',
-		pending: 'To do'
-	};
-
-	function fmtDate(ts?: string | null): string {
-		return ts ? new Date(ts).toLocaleDateString() : '—';
 	}
 
 	// An ended session (submitted or terminated) can never be retaken — even if the
